Move VS Code language service modules into their own folder
