+++ vscode/src/language-service/diagnostics.ts
@@ -1,0 +1,49 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import {
+  ILanguageService,
+  VSDiagnostic,
+  qsharpLibraryUriScheme,
+} from "qsharp-lang";
+import * as vscode from "vscode";
+import { qsharpLanguageId, toVsCodeDiagnostic } from "../common";
+
+export function startLanguageServiceDiagnostics(
+  languageService: ILanguageService,
+): vscode.Disposable[] {
+  const diagCollection =
+    vscode.languages.createDiagnosticCollection(qsharpLanguageId);
+
+  async function onDiagnostics(evt: {
+    detail: {
+      uri: string;
+      version: number;
+      diagnostics: VSDiagnostic[];
+    };
+  }) {
+    const diagnostics = evt.detail;
+    const uri = vscode.Uri.parse(diagnostics.uri);
+
+    if (uri.scheme === qsharpLibraryUriScheme) {
+      // Don't report diagnostics for library files.
+      return;
+    }
+
+    diagCollection.set(
+      uri,
+      diagnostics.diagnostics.map((d) => toVsCodeDiagnostic(d)),
+    );
+  }
+
+  languageService.addEventListener("diagnostics", onDiagnostics);
+
+  return [
+    {
+      dispose: () => {
+        languageService.removeEventListener("diagnostics", onDiagnostics);
+      },
+    },
+    diagCollection,
+  ];
+}
